Allow mostly-complete --fix-includes builds on Linux 2.6.31.
ToDevice and FromHost don't compile yet.

- Fix ASM alternatives.
- Get around some other changed APIs.
- Change "swap" to "linux_swap" in header files.
- The LINUXINCLUDE symbol is used for header files.
- Use cxxprotect where it's required.
- Complex C initializers are getting deeper and deeper into the headers.
  Crap.  Here have to do something about [0...31] = ~0UL.

+++ include/click/crc32.h
@@ -5,9 +5,8 @@
 extern "C" {
 #endif
 
-unsigned long update_crc(unsigned long crc_accum,
-                         const char *data_blk_ptr,
-                         int data_blk_size);
+uint32_t update_crc(uint32_t crc_accum, const char *data_blk_ptr,
+		    int data_blk_size);
 
 #ifdef __cplusplus
 }
